Move JS source and test files into respective folders.
